# packages/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ninotek F&B POS API
  version: 1.0.0
  description: |
    Hợp đồng REST cho toàn hệ thống Ninotek F&B POS.

    **Đây là nguồn sự thật duy nhất cho API.** Client C# (NinoPOS, qua NSwag) và
    Dart (NinoOrder/NinoDash, qua openapi-generator) đều sinh code từ file này.
    CI sẽ fail nếu implementation lệch khỏi contract.

    ## Hai mặt phẳng API tách biệt

    | | LAN API | Cloud API |
    |---|---|---|
    | Máy chủ | NinoPOS tự host (Kestrel) | NinoCloud (NestJS) |
    | Địa chỉ | `http://<ip-máy-POS>:8080` | `https://api.ninotek.vn` |
    | Client | NinoOrder (tablet nhân viên) | NinoPOS (sync), NinoDash |
    | Xác thực | Device token cấp khi ghép nối | JWT Bearer |
    | Yêu cầu Internet | **Không** — chạy hoàn toàn nội bộ | Có |

    Khi mất Internet, LAN API vẫn phục vụ bình thường. Đây là nền tảng của
    kiến trúc Offline-First.

    ## Optimistic Locking

    Mọi endpoint sửa dữ liệu đều nhận `rowVersion` mà client đọc được lúc đầu.
    Nếu bản ghi đã bị người khác sửa, server trả **409 Conflict** kèm bản ghi
    hiện tại để client hiển thị và cho người dùng quyết định.

  contact:
    name: Ninotek — Công nghệ & Giải pháp số
    url: https://ninotek.vn
  license:
    name: Proprietary
    url: https://ninotek.vn/license
servers:
  - url: http://{posHost}:8080
    description: LAN API — NinoPOS tự host trong mạng nội bộ quán
    variables:
      posHost:
        default: 192.168.1.10
...
  - url: https://api.ninotek.vn
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/lan/discovery:
    get:
      tags: [LAN — Thiết bị]
      summary: Thông tin máy POS (dùng để tablet xác nhận đã tìm đúng máy)
      description: |
        Endpoint duy nhất KHÔNG cần xác thực. NinoOrder gọi sau khi dò mDNS
        `_ninopos._tcp.local` để xác nhận đúng máy trước khi ghép nối.
      security: []
      operationId: lanDiscovery
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [storeName, storeCode, posVersion, protocolVersion]
                properties:
                  storeName: { type: string, example: Ninotek Coffee - Quy Nhon }
                  storeCode: { type: string, example: QN01 }
                  posVersion: { type: string, example: 1.0.0 }
                  protocolVersion:
                    type: integer
                    example: 1
                    description: Tablet có protocolVersion khác PHẢI báo người dùng cập nhật app.
                  requiresPairing: { type: boolean }
        '403':
          description: Máy POS đang tắt chế độ cho phép ghép nối thiết bị mới
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/lan/devices/pair:
    post:
      tags: [LAN — Thiết bị]
      summary: Ghép nối tablet với máy POS
      description: |
        Thu ngân phải bấm duyệt trên màn hình NinoPOS thì thiết bị mới được cấp
        token. Số thiết bị ghép nối đồng thời bị giới hạn bởi `licenses.max_devices`.
      security: []
      operationId: lanPairDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceName, deviceType, pairingCode]
              properties:
                deviceName: { type: string, example: Tablet Tầng 1 }
                deviceType: { type: string, enum: [TABLET, PHONE, KDS] }
                pairingCode:
                  type: string
                  description: Mã 6 số hiển thị trên màn hình NinoPOS
                  pattern: '^\d{6}$'
      responses:
        '200':
          description: Ghép nối thành công
          content:
            application/json:
              schema:
                type: object
                required: [deviceId, deviceToken]
                properties:
                  deviceId: { type: string, format: uuid }
                  deviceToken: { type: string, description: Gửi kèm header Authorization ở mọi lời gọi LAN sau đó }
        '403':
          description: Sai mã ghép nối hoặc đã đạt giới hạn thiết bị của license
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/lan/tables/status:
    get:
      tags: [LAN — Bàn]
      summary: Sơ đồ bàn thời gian thực
      operationId: lanGetTableStatus
      parameters:
        - name: areaId
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TableStatus' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/lan/tables/{tableId}/lock:
    post:
      tags: [LAN — Bàn]
      summary: Chiếm quyền thao tác bàn (soft-lock 30 giây)
      description: |
        Chống hai nhân viên cùng sửa một bàn. Lock tự hết hạn sau 30 giây để
        không bị kẹt khi tablet mất kết nối giữa chừng. Gọi lại để gia hạn.
      operationId: lanLockTable
      parameters:
        - $ref: '#/components/parameters/TableId'
      responses:
        '200':
          description: Đã chiếm lock
          content:
            application/json:
              schema:
                type: object
                properties:
                  lockedUntil: { type: string, format: date-time }
        '409':
          description: Bàn đang bị nhân viên khác giữ
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/lan/order/create:
    post:
      tags: [LAN — Order]
      summary: Gửi order từ tablet về máy thu ngân
      description: |
        Đường đi nóng nhất của hệ thống — mục tiêu độ trễ **< 0.5 giây**.

        `orderId` và mọi `orderDetailId` do **client sinh** (UUID v4) trước khi
        gửi. Nhờ vậy khi mất LAN giữa chừng, tablet đệm đơn cục bộ rồi gửi lại
        với cùng UUID; server nhận diện trùng và bỏ qua thay vì tạo đơn thứ hai.
        Đây chính là khoá idempotency của endpoint này.
      operationId: lanCreateOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrderRequest' }
      responses:
        '201':
          description: Đã tạo đơn và đẩy lệnh in xuống bếp
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '200':
          description: Đơn đã tồn tại (client gửi lại sau khi mất kết nối) — trả về đơn cũ, không tạo mới
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          description: Món đã hết hoặc bàn không ở trạng thái nhận order được
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /api/v1/lan/orders/{orderId}:
    get:
      tags: [LAN — Order]
      summary: Chi tiết một đơn hàng
      operationId: lanGetOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/lan/orders/{orderId}/transfer:
    post:
      tags: [LAN — Order]
      summary: Chuyển bàn / gộp bàn
      operationId: lanTransferOrder
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetTableId, rowVersion]
              properties:
                targetTableId: { type: string, format: uuid }
                mergeIfOccupied:
                  type: boolean
                  default: false
                  description: true = gộp vào đơn đang mở ở bàn đích; false = từ chối nếu bàn đã có khách
                rowVersion: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/lan/order-details/{orderDetailId}/kitchen-status:
    patch:
      tags: [LAN — Bếp]
      summary: Cập nhật trạng thái chế biến món (từ màn hình KDS)
      operationId: lanUpdateKitchenStatus
      parameters:
        - name: orderDetailId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kitchenStatus, rowVersion]
              properties:
                kitchenStatus: { $ref: '#/components/schemas/KitchenStatus' }
                rowVersion: { type: integer }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDetail' }
        '409':
          $ref: '#/components/responses/Conflict'
  /api/v1/lan/menu:
    get:
      tags: [LAN — Order]
      summary: Toàn bộ thực đơn (danh mục, món, topping)
      description: |
        NinoOrder cache lại kết quả và dùng `If-None-Match` để chỉ tải khi
        thực đơn thực sự thay đổi — quan trọng vì tablet ở quán thường dùng
        Wi-Fi yếu.
      operationId: lanGetMenu
      parameters:
        - name: If-None-Match
          in: header
          schema: { type: string }
      responses:
        '200':
          description: OK
          headers:
            ETag:
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuSnapshot' }
        '304':
          description: Thực đơn chưa đổi — dùng bản cache
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================================================================
  # CLOUD API
  # ==================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT ngắn hạn (15 phút) cho Cloud API; device token cho LAN API.
    WebhookSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: HMAC-SHA256 của raw body với secret chia sẻ với cổng thanh toán.
  parameters:
    TableId:
      name: tableId
      in: path
      required: true
      schema: { type: string, format: uuid }
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Thiếu hoặc sai thông tin xác thực
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Không tìm thấy
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: |
        Optimistic locking thất bại — bản ghi đã bị người khác sửa.
        Body chứa `currentRowVersion` và bản ghi mới nhất để client hiển thị.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ConflictProblem' }
  schemas:
    Problem:
      type: object
      description: Lỗi theo chuẩn RFC 9457 (Problem Details for HTTP APIs).
      required: [type, title, status]
      properties:
        type: { type: string, format: uri, example: 'https://ninotek.vn/errors/table-locked' }
        title: { type: string, example: Bàn đang bị nhân viên khác thao tác }
        status: { type: integer, example: 409 }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
...
    TableStatus:
      type: object
      required: [id, name, areaId, status, rowVersion]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Bàn 01 }
        areaId: { type: string, format: uuid }
        areaName: { type: string }
        seatCapacity: { type: integer }
        status: { type: string, enum: [EMPTY, OCCUPIED, RESERVED, BILLING, LOCKED] }
        currentOrderId: { type: [string, 'null'], format: uuid }
        guestCount: { type: [integer, 'null'] }
        runningTotal: { type: [number, 'null'], description: 'Tạm tính, VND' }
        seatedMinutes: { type: [integer, 'null'] }
        lockedByUserId: { type: [string, 'null'], format: uuid }
        lockedUntil: { type: [string, 'null'], format: date-time }
        posX: { type: integer }
        posY: { type: integer }
        rowVersion: { type: integer }
    CreateOrderRequest:
      type: object
      required: [orderId, userId, orderType, items]
      properties:
        orderId:
          type: string
          format: uuid
          description: UUID v4 do CLIENT sinh — đồng thời là khoá idempotency khi gửi lại.
        tableId: { type: [string, 'null'], format: uuid }
        userId: { type: string, format: uuid }
        orderType: { type: string, enum: [DINE_IN, TAKE_AWAY, DELIVERY], default: DINE_IN }
        guestCount: { type: integer, minimum: 1, default: 1 }
        note: { type: string }
        clientCreatedAt:
          type: string
          format: date-time
          description: Thời điểm trên tablet. Giúp giữ đúng thứ tự khi nhiều đơn được đệm offline rồi gửi cùng lúc.
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [orderDetailId, itemId, quantity]
            properties:
              orderDetailId: { type: string, format: uuid, description: UUID v4 do client sinh }
              itemId: { type: string, format: uuid }
              quantity: { type: number, minimum: 0.01 }
              note: { type: string, example: 'ít đường, không đá' }
              toppings:
                type: array
                items:
                  type: object
                  required: [toppingId]
                  properties:
                    toppingId: { type: string, format: uuid }
                    quantity: { type: integer, minimum: 1, default: 1 }
    Order:
      type: object
      required: [id, orderCode, status, subtotal, finalTotal, rowVersion]
      properties:
        id: { type: string, format: uuid }
        orderCode: { type: string, example: NINOPOS100234 }
        tableId: { type: [string, 'null'], format: uuid }
        tableName: { type: [string, 'null'] }
        shiftId: { type: [string, 'null'], format: uuid }
        userId: { type: string, format: uuid }
        orderType: { type: string, enum: [DINE_IN, TAKE_AWAY, DELIVERY] }
        status: { type: string, enum: [DRAFT, PENDING, SERVING, COMPLETED, CANCELLED] }
        guestCount: { type: integer }
        subtotal: { type: number }
        discountAmount: { type: number }
        serviceFee: { type: number }
        vatAmount: { type: number }
        finalTotal: { type: number }
        paidAmount: { type: number }
        note: { type: [string, 'null'] }
        details:
          type: array
          items: { $ref: '#/components/schemas/OrderDetail' }
        createdAt: { type: string, format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        updatedAt: { type: string, format: date-time }
        rowVersion: { type: integer }
...